put stuff from IntegratedPlugins to Mining.Plugins namespace
