+++ src/NiceHashMiner/Algorithms/PluginAlgorithm.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using NiceHashMiner.Plugin;
+using NiceHashMiner.Mining.Plugins;
 using NHM.Common.Enums;
 using AlgorithmCommon = NHM.Common.Algorithm;
