+++ src/NiceHashMiner/ApplicationStateManager/ApplicationStateManager.Init.cs
@@ -2,8 +2,7 @@
 using NiceHashMiner.Configs;
 using NiceHashMiner.Devices;
 using NiceHashMiner.Forms.Components;
-using NiceHashMiner.Mining.IntegratedPlugins;
-using NiceHashMiner.Plugin;
+using NiceHashMiner.Mining.Plugins;
 using NiceHashMiner.Stats;
 using NiceHashMiner.Utils;
 using NHM.Common;
