+++ src/NiceHashMiner/Mining/Grouping/GroupingUtils.cs
@@ -1,6 +1,6 @@
 ﻿using NHM.Common.Enums;
 using NiceHashMiner.Algorithms;
-using NiceHashMiner.Plugin;
+using NiceHashMiner.Mining.Plugins;
 using System.Collections.Generic;
 using System.Linq;
 using MinerPlugin;
