--- src/NiceHashMiner/Mining/IntegratedPlugins/IntegratedPlugin.cs
@@ -1,9 +1,0 @@
-﻿using MinerPlugin;
-
-namespace NiceHashMiner.Mining.IntegratedPlugins
-{
-    public interface IntegratedPlugin : IMinerPlugin
-    {
-        bool Is3rdParty { get; }
-    }
-}
