+++ src/NiceHashMiner/Mining/Plugins/IntegratedPlugins/CCMinerIntegrated/CCMinerMTPIntegratedPlugin.cs
@@ -11,7 +11,7 @@ using NHM.Common.Algorithm;
 using NHM.Common.Device;
 using NHM.Common.Enums;
 
-namespace NiceHashMiner.Mining.IntegratedPlugins
+namespace NiceHashMiner.Mining.Plugins
 {
     class CCMinerMTPIntegratedPlugin : CCMinersPluginBase
     {
