+++ src/NiceHashMiner/Mining/Plugins/IntegratedPlugins/SGminerIntegratedHub/SGminerIntegratedMiner.cs
@@ -10,7 +10,7 @@ using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace NiceHashMiner.Mining.IntegratedPlugins
+namespace NiceHashMiner.Mining.Plugins
 {
     class SGminerIntegratedMiner : SGMinerBase
     {
