+++ src/NiceHashMiner/NiceHashMiner.csproj
@@ -242,26 +242,25 @@
     <Compile Include="Mining\IdleChecking\IdleCheckManager.cs" />
     <Compile Include="Mining\IdleChecking\InputTimeoutChecker.cs" />
     <Compile Include="Mining\IdleChecking\SessionLockChecker.cs" />
-    <Compile Include="Mining\IntegratedPlugins\CCMinerIntegrated\CCMinerIntegratedMiner.cs" />
-    <Compile Include="Mining\IntegratedPlugins\CCMinerIntegrated\CCMinerMTPIntegratedPlugin.cs" />
-    <Compile Include="Mining\IntegratedPlugins\CCMinerIntegrated\CCMinersPluginBase.cs" />
-    <Compile Include="Mining\IntegratedPlugins\CCMinerIntegrated\CCMinerTpruvotIntegratedPlugin.cs" />
-    <Compile Include="Mining\IntegratedPlugins\IntegratedPlugin.cs" />
-    <Compile Include="Mining\IntegratedPlugins\IntegratedPluginClasses.cs" />
-    <Compile Include="Mining\IntegratedPlugins\MinersBinsUrls.cs" />
-    <Compile Include="Mining\IntegratedPlugins\SGminerIntegratedHub\SGminerAvemoreIntegratedPlugin.cs" />
-    <Compile Include="Mining\IntegratedPlugins\SGminerIntegratedHub\SGminerGMIntegratedPlugin.cs" />
-    <Compile Include="Mining\IntegratedPlugins\SGminerIntegratedHub\SGminerIntegratedMiner.cs" />
-    <Compile Include="Mining\IntegratedPlugins\SGminerIntegratedHub\SGMinerPluginBase.cs" />
-    <Compile Include="Mining\IntegratedPlugins\VC_REDIST_x64_2015_DEPENDENCY_PLUGIN.cs" />
+    <Compile Include="Mining\Plugins\IntegratedPlugins\CCMinerIntegrated\CCMinerIntegratedMiner.cs" />
+    <Compile Include="Mining\Plugins\IntegratedPlugins\CCMinerIntegrated\CCMinerMTPIntegratedPlugin.cs" />
+    <Compile Include="Mining\Plugins\IntegratedPlugins\CCMinerIntegrated\CCMinersPluginBase.cs" />
+    <Compile Include="Mining\Plugins\IntegratedPlugins\CCMinerIntegrated\CCMinerTpruvotIntegratedPlugin.cs" />
+    <Compile Include="Mining\Plugins\IntegratedPlugins.cs" />
+    <Compile Include="Mining\Plugins\MinersBinsUrls.cs" />
+    <Compile Include="Mining\Plugins\IntegratedPlugins\SGminerIntegratedHub\SGminerAvemoreIntegratedPlugin.cs" />
+    <Compile Include="Mining\Plugins\IntegratedPlugins\SGminerIntegratedHub\SGminerGMIntegratedPlugin.cs" />
+    <Compile Include="Mining\Plugins\IntegratedPlugins\SGminerIntegratedHub\SGminerIntegratedMiner.cs" />
+    <Compile Include="Mining\Plugins\IntegratedPlugins\SGminerIntegratedHub\SGMinerPluginBase.cs" />
+    <Compile Include="Mining\Plugins\VC_REDIST_x64_2015_DEPENDENCY_PLUGIN.cs" />
     <Compile Include="Mining\MinerApiWatchdog.cs" />
     <Compile Include="Mining\MiningManager.cs" />
     <Compile Include="Mining\MiningStats.cs" />
     <Compile Include="NHMProductInfo.cs" />
-    <Compile Include="Plugin\PluginContainer.cs" />
-    <Compile Include="Plugin\SupportedAlgorithmsFilter.cs" />
-    <Compile Include="Plugin\SupportedPluginsFilter.cs" />
-    <Compile Include="Plugin\MinerPluginsManager.cs" />
+    <Compile Include="Mining\Plugins\PluginContainer.cs" />
+    <Compile Include="Mining\Plugins\SupportedAlgorithmsFilter.cs" />
+    <Compile Include="Mining\Plugins\SupportedPluginsFilter.cs" />
+    <Compile Include="Mining\Plugins\MinerPluginsManager.cs" />
     <Compile Include="Mining\MinerFromPlugin.cs" />
     <Compile Include="Stats\ErrorCode.cs" />
     <Compile Include="Stats\Models\ExchangeRateJson.cs" />
@@ -275,8 +274,8 @@
     <Compile Include="Stats\Models\LoginMessage.cs" />
     <Compile Include="Stats\Models\DeviceStatusMessage.cs" />
     <Compile Include="Stats\Models\RpcException.cs" />
-    <Compile Include="Plugin\PluginPackageInfo.cs" />
-    <Compile Include="Plugin\PluginPackageInfoCR.cs" />
+    <Compile Include="Mining\Plugins\PluginPackageInfo.cs" />
+    <Compile Include="Mining\Plugins\PluginPackageInfoCR.cs" />
     <Compile Include="Stats\ExchangeRateAPI.cs" />
     <Compile Include="Stats\NiceHashSocket.cs" />
     <Compile Include="Stats\NiceHashStats_shared.cs" />
