+++ src/NiceHashMiner/Utils/FirewallRules.cs
@@ -5,6 +5,7 @@ using System.Diagnostics;
 using System.IO;
 using System.Linq;
 using MinerPluginToolkitV1.Configs;
+using NiceHashMiner.Mining.Plugins;
 
 namespace NiceHashMiner.Utils
 {
@@ -22,7 +23,7 @@ namespace NiceHashMiner.Utils
 
         public static bool IsFirewallRulesOutdated()
         {
-            var installedPlugins = Plugin.MinerPluginsManager.GetPluginUUIDsAndVersionsList();
+            var installedPlugins = MinerPluginsManager.GetPluginUUIDsAndVersionsList();
             return installedPlugins.Except(_pluginsUUIDsWithVersions).Count() > 0;
         }
 
@@ -50,7 +51,7 @@ namespace NiceHashMiner.Utils
                     else
                     {
                         Logger.Info("NICEHASH", "setFirewallRulesProcess all OK");
-                        var installedPlugins = Plugin.MinerPluginsManager.GetPluginUUIDsAndVersionsList();
+                        var installedPlugins = MinerPluginsManager.GetPluginUUIDsAndVersionsList();
                         _pluginsUUIDsWithVersions = installedPlugins;
                         InternalConfigs.WriteFileSettings(_firewallRulesAddedFilePath, _pluginsUUIDsWithVersions);
                     }
